Move command to seperate file and fixed ticket sale with update button
