+++ databases/ticket_sales_db.txt
@@ -6,3 +6,6 @@ Screen 4 -  Gattaca,61,
 Screen 7 -  Top Secret,59,
 Screen 9 -  Midsommar,101,
 Screen 7 - Seven ,1,
+Screen 6 - AVATAR ,1,
+Screen 2 - ERASERHEAD ,1,
+Screen 4 - DREDD ,1,
