+++ logic_pkg/entry_logic.py
@@ -2,75 +2,8 @@ from controllers import ui_controller, logic_controller
 from utils import utils
 from interceptors import encryption_interceptor
 import requests
-import abc
+from entities import command
 
-class Invoker:
-    # sends request to command
-    def __init__(self):
-        self._commands = []
-
-    def store_command(self, command):
-        self._commands.append(command)
-
-    def execute_commands(self):
-        for command in self._commands:
-            command.execute()
-
-class Command(metaclass=abc.ABCMeta):
-    #interface for all concrete commands
-    def __init__(self, login):
-        self._login = login
-
-    @abc.abstractmethod
-    def execute(self):
-        pass
-
-
-    #Define a binding between a Receiver object and an action.
-    #Implement Execute by invoking the corresponding operation(s) on
-    #Receiver. 
-class UserLoginCommand(Command) : 
-
-    def __init__(self, receiver, userName):
-        self._receiver = receiver
-        self._userName = userName
-
-
-    def execute(self) :
-        self._receiver.user_login(self._userName)
-
-
-class AdminLoginCommand(Command) : 
-
-    def __init__(self, receiver, userName):
-        self._receiver = receiver
-        self._userName = userName
-
-
-    def execute(self) :
-        self._receiver.admin_login(self._userName)
-    
-
-#Reciever below
-class Receiver :
-    
-    def user_login(self, userName):
-        ui_controller.ui.get_current_ui().Hide()
-        ui_controller.ui.open_main_menu_user_ui()
-        ui_controller.ui.set_current_user(userName)
-        logic_controller.logic.set_current_user(userName)
-        logic_controller.logic.set_main_menu_user_loop()
-        logic_controller.logic.set_auth_type("user")
-
-    def admin_login(self, userName):
-        ui_controller.ui.get_current_ui().Hide()
-        ui_controller.ui.open_main_menu_admin_ui()
-        ui_controller.ui.set_current_user(userName)
-        logic_controller.logic.set_current_user(userName)
-        logic_controller.logic.set_main_menu_admin_loop()
-        logic_controller.logic.set_auth_type("admin")  
-
-    
 
 def loginEventLoop(window, event, values):
     window['-USERNAME-'].update("")
@@ -128,16 +61,16 @@ def validateLogin(window, username, password):
     usernames = loginData.readline().split(",")
     passwords = loginData.readline().split(",")
     authTypes = loginData.readline().split(",")
-    receiver = Receiver()
-    invoker = Invoker()
+    invoker = command.Invoker()
+    receiver = command.Receiver()
     if username in usernames:
         if password == utils.decrypt(passwords[usernames.index(username)]):
             if authTypes[usernames.index(username)] == "admin":
-                admin_login = AdminLoginCommand(receiver, username)
+                admin_login = command.AdminLoginCommand(receiver, username)
                 invoker.store_command(admin_login)
                 invoker.execute_commands()
             else:
-                user_login = UserLoginCommand(receiver, username)
+                user_login = command.UserLoginCommand(receiver, username)
                 invoker.store_command(user_login)
                 invoker.execute_commands()
         else:
