+++ ui_pkg/ticket_sales_ui.py
@@ -11,4 +11,4 @@ ticket_sales_screen = ticket_sales_list.generate_list()
 
 adminLayout = [[sg.Text("TheAter Ticket Sales")], 
              [sg.Listbox(ticket_sales_screen, size=(100, len(ticket_sales_screen)), key='-List-', enable_events=True)],
-             [sg.Button('Back To Menu')]]
+             [sg.Button('Update Values'), sg.Button('Back To Menu')]]
